Functional sidebar and number simplifier
Replaced sidebar text with functional variables that update when the player clicks on a country, includes country name, population, GDP, science, GDP per capita, and military power. Removed tooltip, as it was unnecessary with the sidebar displaying country details. In addition, I added a function that makes numbers readability of details much easier by changing it to thousands, millions, billions and trillions instead of the raw number. Lastly, the sidebar has helped me identify a bug in the game, which is that some of the countries have not been shown on the map due to issues in the world.geojson and countries.json file. Where the country code, (which is used to identify the country and process the details) being broken, as the country code had been put as -99 instead of the actual code. But, will discuss this in more depth in the dev log.