MiniWeb added!
This took way to long to incorporate, with accidently deleting my project and breaking it in several ways to try to get this extension to work, it finally does work! This means that anyone who downloads the ZIP folder of my game can play it offline, without downloading any additional extensions!