Game menu added before game screen, changed colour theme to match global command logo, player country selection development.
The player now enters the newly created main menu before going straight to the game. I also changed the colours to better match the global command logo. I also refined the boarder edges so everywhere has a boarder with the same width and no overlapping, which made some of the boarders thicker. Lastly, I have started working on the player country selection system, which will "lock in" players country choice, and later let them play as the chosen country. For now, the changes I have made are that the new country selection button can be clicked when the player has clicked on a country, which will now turn white until the user decides to look at a different country or lock the country in and play as the country selected (which is what I am working on now).