Starting gameplay mode and working select country button
The select country button now works, where the player can click on a country, and set the country as their chosen country when clicking on the select country button. Along with this, the sidebar is rebuilt, serving as a skeleton for the gameplay UI, instead of the a country preview/interactive map sidebar. It now displays time speeds and pause. In addition, there is a settings button and area to be filled out with possible country decisions on GDP, population, military, etc. Lastly, there is a diplomacy part of the sidebar which enables the player to still be able to view other countries and their details. None of these are functional at the moment, but will be in future updates.