End of sprint 2, JS scripts rework and functional time controls.
This update will mark the end of sprint 2.     In this final update for sprint 2, the main.JS code has been broken up into six new, JS files which all have important and clear functions (map.js, main.js, sidebar.js, states.js, timer.js, and utlilities.js). The files work together through imports and exports. This has allowed for it to be much easier to find parts of code, when fixing trouble shooting. In addition to this, I have also made a fully functional time system, with the ability to pause and increase time speed through 1x, 3x or 10x buttons, with a much improved UI layout as well. I believe the game is now developed enough to move on to sprint 3, which can be fully focused on gameplay.